refactor: move prompts to own store/reactquery modules, add in filter WIP
