+++ client/src/components/Prompts/Groups/GroupSidePanel.tsx
@@ -1,17 +1,18 @@
-import { useMemo } from 'react';
+import { useMemo, useState } from 'react';
+import { ListFilter } from 'lucide-react';
 import { useLocation } from 'react-router-dom';
-// import { ListFilter } from 'lucide-react';
-// import { useNavigate, useLocation } from 'react-router-dom';
 import PanelNavigation from '~/components/Prompts/Groups/PanelNavigation';
 import { useMediaQuery, usePromptGroupsNav } from '~/hooks';
 import List from '~/components/Prompts/Groups/List';
-// import { Button, Input } from '~/components/ui';
+import { Button, Input } from '~/components/ui';
 import { cn } from '~/utils';
 
 export default function GroupSidePanel({
   children,
   isDetailView,
   className = '',
+  /* usePromptGroupsNav */
+  setName,
   nextPage,
   prevPage,
   isFetching,
@@ -24,6 +25,7 @@ export default function GroupSidePanel({
   className?: string;
 } & ReturnType<typeof usePromptGroupsNav>) {
   const location = useLocation();
+  const [displayName, setDisplayName] = useState('');
   const isSmallerScreen = useMediaQuery('(max-width: 1024px)');
   const isChatRoute = useMemo(() => location.pathname.startsWith('/c/'), [location.pathname]);
 
@@ -35,21 +37,24 @@ export default function GroupSidePanel({
         className,
       )}
     >
-      <div className="flex w-full flex-row justify-between">
-        {/* <div className="mx-4 flex w-2/3 flex-row justify-start gap-x-2 pr-2">
-          <Button variant="ghost" className="m-0 mr-2 p-0">
-          <ListFilter className="h-4 w-4" />
-          </Button>
-          <Input
-          placeholder='Filter prompts...'
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-          className="max-w-sm dark:border-gray-500"
-          />
-          </div> */}
-      </div>
       {children}
       <div className="flex-grow overflow-y-auto">
+        <div className="m-2 flex w-full flex-row justify-between">
+          <div className="flex">
+            <Button variant="outline" size="sm" className="h-10 w-10">
+              <ListFilter className="icon-sm" />
+            </Button>
+            <Input
+              placeholder="Filter prompts..."
+              value={displayName}
+              onChange={(e) => {
+                setDisplayName(e.target.value);
+                setName(e.target.value);
+              }}
+              className="max-w-sm dark:border-gray-500"
+            />
+          </div>
+        </div>
         <List groups={promptGroups} isChatRoute={isChatRoute} />
       </div>
       <PanelNavigation
