+++ client/src/components/Prompts/PreviewLabels.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Cross1Icon } from '@radix-ui/react-icons';
 import type { TPrompt } from 'librechat-data-provider';
-import { useUpdatePromptLabels } from '~/data-provider/mutations';
+import { useUpdatePromptLabels } from '~/data-provider';
 import { Input } from '~/components/ui';
 
 const PromptForm = ({ selectedPrompt }: { selectedPrompt?: TPrompt }) => {
