+++ client/src/data-provider/index.ts
@@ -1,3 +1,4 @@
 export * from './mutations';
+export * from './prompts';
 export * from './queries';
 export * from './roles';
