+++ client/src/store/index.ts
@@ -8,6 +8,7 @@ import toast from './toast';
 import submission from './submission';
 import search from './search';
 import preset from './preset';
+import prompts from './prompts';
 import lang from './language';
 import settings from './settings';
 
@@ -21,6 +22,7 @@ export default {
   ...toast,
   ...submission,
   ...search,
+  ...prompts,
   ...preset,
   ...lang,
   ...settings,
