+++ client/src/store/prompts.ts
@@ -1,0 +1,20 @@
+import { atom } from 'recoil';
+import { atomWithLocalStorage } from '~/store/utils';
+
+// Static atoms without localStorage
+const staticAtoms = {
+  // `name` filter
+  promptsName: atom<string>({ key: 'promptsName', default: '' }),
+  // `pageNumber` filter
+  promptsPageNumber: atom<number>({ key: 'promptsPageNumber', default: 1 }),
+  // `pageSize` filter
+  promptsPageSize: atom<number>({ key: 'promptsPageSize', default: 10 }),
+};
+
+// Atoms with localStorage
+const localStorageAtoms = {
+  autoSendPrompts: atomWithLocalStorage('autoSendPrompts', true),
+  alwaysMakeProd: atomWithLocalStorage('alwaysMakeProd', true),
+};
+
+export default { ...staticAtoms, ...localStorageAtoms };
