+++ client/src/store/settings.ts
@@ -1,36 +1,7 @@
 import { atom } from 'recoil';
 import { SettingsViews } from 'librechat-data-provider';
+import { atomWithLocalStorage } from '~/store/utils';
 import type { TOptionSettings } from '~/common';
-
-// Improved helper function to create atoms with localStorage
-function atomWithLocalStorage<T>(key: string, defaultValue: T) {
-  return atom<T>({
-    key,
-    default: defaultValue,
-    effects_UNSTABLE: [
-      ({ setSelf, onSet }) => {
-        const savedValue = localStorage.getItem(key);
-        if (savedValue !== null) {
-          try {
-            const parsedValue = JSON.parse(savedValue);
-            setSelf(parsedValue);
-          } catch (e) {
-            console.error(
-              `Error parsing localStorage key "${key}", \`savedValue\`: defaultValue, error:`,
-              e,
-            );
-            localStorage.setItem(key, JSON.stringify(defaultValue));
-            setSelf(defaultValue);
-          }
-        }
-
-        onSet((newValue: T) => {
-          localStorage.setItem(key, JSON.stringify(newValue));
-        });
-      },
-    ],
-  });
-}
 
 // Static atoms without localStorage
 const staticAtoms = {
@@ -73,8 +44,6 @@ const localStorageAtoms = {
   rememberForkOption: atomWithLocalStorage('rememberForkOption', true),
   playbackRate: atomWithLocalStorage<number | null>('playbackRate', null),
   saveDrafts: atomWithLocalStorage('saveDrafts', false),
-  autoSendPrompts: atomWithLocalStorage('autoSendPrompts', true),
-  alwaysMakeProd: atomWithLocalStorage('alwaysMakeProd', true),
 };
 
 export default { ...staticAtoms, ...localStorageAtoms };
